% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buttons.R
\name{SuccessButton}
\alias{SuccessButton}
\title{Success Button Component}
\usage{
SuccessButton(inputId, label, icon = NULL, width = NULL, ...)
}
\arguments{
\item{inputId}{The input ID of the button.}

\item{label}{The text to display on the button.}

\item{icon}{An optional icon for the button.}

\item{width}{An optional width for the button.}

\item{...}{Additional arguments passed to the actionButton.}
}
\value{
A styled success button for Shiny apps.
}
\description{
Success Button Component
}
